+++ bw_temporalis/__init__.py
@@ -3,7 +3,7 @@ __all__ = (
     "check_database_exchanges",
     "easy_datetime_distribution",
     "easy_timedelta_distribution",
-    "FixedTimeOfYear",
+    "FixedTimeOfYearTD",
     "FixedTD",
     "IncongruentDistribution",
     "loader_registry",
@@ -15,8 +15,12 @@ __all__ = (
 )
 
 
-from .base_dynamic_class import TDAware
-from .temporal_distribution import TemporalDistribution
+from .temporal_distribution import (
+    TemporalDistribution,
+    FixedTimeOfYearTD,
+    FixedTD,
+    TDAware,
+)
 from .timeline import Timeline
 from .lca import TemporalisLCA
 from .utils import (
@@ -27,11 +31,11 @@ from .utils import (
     get_version_tuple,
     supplement_dataframe,
 )
-from .example_functions import FixedTimeOfYear, FixedTD
 
 loader_registry = {
-    "bw_temporalis.temporal_distribution.TemporalDistribution.from_json": TemporalDistribution.from_json,
-    "bw_temporalis.example_functions.FixedTimeOfYear.from_json": FixedTimeOfYear.from_json,
+    "bw_temporalis.TemporalDistribution": TemporalDistribution.from_json,
+    "bw_temporalis.FixedTimeOfYear": FixedTimeOfYearTD.from_json,
+    "bw_temporalis.FixedTD": FixedTD.from_json,
 }
 
 __version__ = get_version_tuple()