+++ bw_temporalis/temporal_distribution.py
@@ -1,13 +1,12 @@
 import json
 from collections.abc import Mapping
 from numbers import Number
-from typing import Any, SupportsFloat, Union
+from typing import Any, Optional, SupportsFloat, Union
 
 import numpy as np
 import numpy.typing as npt
 from scipy.cluster.vq import kmeans2
 
-from . import TDAware
 from .convolution import (
     consolidate,
     datetime_type,
@@ -27,7 +26,70 @@ RESOLUTION_LABELS = {
 }
 
 
-class TemporalDistribution:
+class TDAware:
+    """Base class for functions which can be multiplied by temporal distributions"""
+
+    def __mul__(self, other):
+        raise NotImplementedError("Must be defined in child classes")
+
+
+class TemporalDistributionBase:
+    """Base class for temporal distributions"""
+
+    _mul_comes_first = False  # Base class doesn't care about order
+
+    def __len__(self) -> int:
+        return self.amount.shape[0]
+
+    def __lt__(self, other: Any) -> bool:
+        if not isinstance(other, TemporalDistributionBase):
+            return False
+        return self.amount.sum() < other.amount.sum()
+
+    def graph(
+        self, style: str | None = "fivethirtyeight", resolution: str | None = None
+    ):
+        """Graph the temporal distribution.
+
+        `resolution` is one of `YMWDhms`.
+
+        This isn't too difficult, if you need more customization write your own :)"""
+        try:
+            from matplotlib import pyplot as plt
+        except ImportError:
+            raise ImportError("`matplotlib` required for this function")
+        plt.style.use(style)
+        axis = plt.gca()
+
+        if np.issubdtype(self.date.dtype, np.datetime64):
+            axis.set_xlabel("Date")
+            date = self.date
+        else:
+            if resolution is None:
+                date = self.date
+                axis.set_xlabel("Time (seconds)")
+            else:
+                date = self.date.astype(f"timedelta64[{resolution}]")
+                axis.set_xlabel("Time ({})".format(RESOLUTION_LABELS[resolution]))
+        axis.plot(date, self.amount, marker=".", lw=0)
+        axis.set_ylabel("Amount")
+        plt.tight_layout()
+        plt.xticks(rotation=30)
+
+        return axis
+
+    def __str__(self) -> str:
+        return "%s instance with %s values and total: %.4g" % (
+            self.__class__.__name__,
+            len(self.amount),
+            self.total,
+        )
+
+    def __repr__(self) -> str:
+        return str(self)
+
+
+class TemporalDistribution(TemporalDistributionBase):
     """A container for a series of amount spread over time.
     Args:
         * *date* (ndarray): 1D array containg temporal info of `amount` with type `timedelta64` or `datetime64` .
@@ -64,9 +126,6 @@ class TemporalDistribution:
         else:
             raise ValueError("`date` must be numpy datetime or timedelta array")
 
-    def __len__(self) -> int:
-        return self.amount.shape[0]
-
     @property
     def total(self) -> float:
         return float(self.amount.sum())
@@ -75,6 +134,8 @@ class TemporalDistribution:
         self, other: Union["TemporalDistribution", SupportsFloat, TDAware]
     ) -> "TemporalDistribution":
         if isinstance(other, TDAware):
+            return other * self
+        elif isinstance(other, TemporalDistributionBase) and other._mul_comes_first:
             return other * self
         elif isinstance(other, TemporalDistribution):
             if (
@@ -120,11 +181,6 @@ class TemporalDistribution:
             raise ValueError("Can only divide time deltas by a number")
         return TemporalDistribution(self.date, self.amount / float(other))
 
-    def __lt__(self, other: Any) -> bool:
-        if not isinstance(other, TemporalDistribution):
-            return False
-        return self.amount.sum() < other.amount.sum()
-
     def __add__(
         self, other: Union["TemporalDistribution", SupportsFloat]
     ) -> "TemporalDistribution":
@@ -157,51 +213,10 @@ class TemporalDistribution:
                 "Can't add TemporalDistribution and {}".format(type(other))
             )
 
-    def __str__(self) -> str:
-        return "TemporalDistribution instance with %s amount and total: %.4g" % (
-            len(self.amount),
-            self.total,
-        )
-
-    def __repr__(self) -> str:
-        return str(self)
-
-    def graph(
-        self, style: str | None = "fivethirtyeight", resolution: str | None = None
-    ):
-        """Graph the temporal distribution.
-
-        `resolution` is one of `YMWDhms`.
-
-        This isn't too difficult, if you need more customization write your own :)"""
-        try:
-            from matplotlib import pyplot as plt
-        except ImportError:
-            raise ImportError("`matplotlib` required for this function")
-        plt.style.use(style)
-        axis = plt.gca()
-
-        if np.issubdtype(self.date.dtype, np.datetime64):
-            axis.set_xlabel("Date")
-            date = self.date
-        else:
-            if resolution is None:
-                date = self.date
-                axis.set_xlabel("Time (seconds)")
-            else:
-                date = self.date.astype(f"timedelta64[{resolution}]")
-                axis.set_xlabel("Time ({})".format(RESOLUTION_LABELS[resolution]))
-        axis.plot(date, self.amount, marker=".", lw=0)
-        axis.set_ylabel("Amount")
-        plt.tight_layout()
-        plt.xticks(rotation=30)
-
-        return axis
-
     def to_json(self):
         return json.dumps(
             {
-                "__loader__": "bw_temporalis.temporal_distribution.TemporalDistribution.from_json",
+                "__loader__": "bw_temporalis.TemporalDistribution",
                 "date_dtype": str(self.date.dtype),
                 "date": self.date.astype(int).tolist(),
                 "amount": self.amount.tolist(),
@@ -209,7 +224,7 @@ class TemporalDistribution:
         )
 
     @classmethod
-    def from_json(cls, json_obj: str | Mapping) -> "TemporalDistribution":
+    def from_json(cls, json_obj: str | Mapping) -> "TemporalDistributionBase":
         if isinstance(json_obj, Mapping):
             data = json_obj
         elif isinstance(json_obj, str):
@@ -276,3 +291,114 @@ class TemporalDistribution:
         # so we need to remove unused clusters.
         date = np.sort(means[np.unique(codebook)]).astype(self.date.dtype)
         return TemporalDistribution(date=date, amount=amount)
+
+
+class FixedTimeOfYearTD(TemporalDistribution):
+    """Instead of creating a relative shift in time, regardless of the starting datetime or the properties of the system being modeled, create an absolute period in the first available year.
+
+    The fixed time of year is a *relative temporal distribution*, and must already be constructed as such.
+
+    When multiplied by another temporal distribution, the entire period of the `FixedTimeOfYear` must lie before the other distribution, or else the `FixedTimeOfYear` is shifted back a year. In other words, a `FixedTimeOfYear` of March to May could be multiplied by August 2020 and result in March to May 2020, but if multiplied by April 2020 it would result in March to May 2019. To allow partial overlaps and stay in the same year, set `allow_overlap` to `True`. In any case, the start of the `FixedTimeOfYear` must be before the start of the temporal distribution being multiplied or it is shifted back a year.
+
+    """
+
+    _mul_comes_first = True
+
+    def __init__(
+        self,
+        date: npt.NDArray[np.timedelta64],
+        amount: npt.NDArray,
+        allow_overlap: Optional[bool] = False,
+    ):
+        if not isinstance(date, np.ndarray) or not isinstance(amount, np.ndarray):
+            raise ValueError("Invalid input types")
+        elif not date.shape == amount.shape:
+            raise ValueError("Shapes of input `date` and `amount` not identical")
+        elif not np.issubdtype(date.dtype, np.timedelta64):
+            raise ValueError(f"Incorrect `date` dtype ({date.dtype})")
+        elif not len(date):
+            raise ValueError("Empty array")
+        self.allow_overlap = allow_overlap
+
+        self.amount = amount.astype(np.float64)
+        self.date = date.astype(timedelta_type)
+        self.base_time_type = timedelta_type
+        if not np.alltrue(self.date.astype(int) >= 0):
+            raise ValueError("Can't have negative relative timedelta64 values")
+        elif not np.alltrue(self.date.astype(int) <= 365 * 24 * 60 * 60):
+            raise ValueError("Can't have timedelta64 values more than one year")
+
+    def __mul__(
+        self, other: TemporalDistribution | Number
+    ) -> TemporalDistribution | TDAware:
+        if isinstance(other, Number):
+            return FixedTimeOfYearTD(
+                date=self.date,
+                amount=self.amount * other,
+                allow_overlap=self.allow_overlap,
+            )
+        elif isinstance(other, TemporalDistribution):
+            if other.base_time_type == timedelta_type:
+                raise ValueError(
+                    "Can't multiply `FixedTimeOfYearTD` by a relative distribution"
+                )
+
+            previous_year_mask = (
+                other.date - other.date.astype("datetime64[Y]").astype("datetime64[s]")
+            ) > (self.date.min() if self.allow_overlap else self.date.max())
+            return TemporalDistribution(
+                date=other.date.astype("datetime64[Y]") - 1 + previous_year_mask,
+                amount=other.amount,
+            ) * TemporalDistribution(date=self.date, amount=self.amount)
+        else:
+            raise ValueError(
+                "Can only be multipled by a number or an instance of `TemporalDistribution`"
+            )
+
+    def to_json(self) -> str:
+        return json.dumps(
+            {
+                "__loader__": "bw_temporalis.FixedTimeOfYearTD",
+                "date_dtype": str(self.date.dtype),
+                "date": self.date.astype(int).tolist(),
+                "amount": self.amount.tolist(),
+                "allow_overlap": self.allow_overlap,
+            }
+        )
+
+    @classmethod
+    def from_json(cls, json_obj):
+        if isinstance(json_obj, Mapping):
+            data = json_obj
+        elif isinstance(json_obj, str):
+            data = json.loads(json_obj)
+        else:
+            raise ValueError(f"Can't understand `from_json` input object {json_obj}")
+        td = TemporalDistribution.from_json(data)
+        return cls(
+            date=td.date,
+            amount=td.amount,
+            allow_overlap=data["allow_overlap"],
+        )
+
+
+class FixedTD(TemporalDistribution):
+    """An absolute `TemporalDistribution` that ignores other temporal information."""
+
+    _mul_comes_first = True
+
+    def __mul__(
+        self, other: TemporalDistribution | Number
+    ) -> TemporalDistribution | TDAware:
+        coeff = other if isinstance(other, Number) else other.amount.sum()
+        return TemporalDistribution(date=self.date, amount=self.amount * coeff)
+
+    def to_json(self) -> str:
+        return json.dumps(
+            {
+                "__loader__": "bw_temporalis.FixedTD",
+                "date_dtype": str(self.date.dtype),
+                "date": self.date.astype(int).tolist(),
+                "amount": self.amount.tolist(),
+            }
+        )